requirements.txt
pytest
flake8
django
django-crispy-forms
django-betterforms
numpy
